# src/main/resources/META-INF/openapi.yaml
openapi: 3.0.1
info:
  title: OpenAPI template quarkus
  description: Documentação para template quarkus
  version: "1.0.0"

servers:
  - url: http://localhost:8080/
    description: Dev

paths:
  /v1/users:
    get:
      description: Buscar todos os usuários
      responses:
        200:
          description: OK
          content:
            application/json: {}
    post:
      description: Criar um novo usuário
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
            example:
              bio: "Bem vindo a minha bio"
              blog: "https://meublog.com"
              data: "2025-11-03"
              location: "Minha cidade"
              login: "marcel.braghini"
              name: "Marcel Braghini"
      responses:
        201:
          description: Created
          content:
            application/json: {}
        400:
          description: Bad Request
          content:
            application/json: {}
  /v1/users/{_id}:
    parameters:
      - name: _id
        in: path
        required: true
        description: Id do usuário
        schema:
          type: string
    get:
      description: Buscar um usuário pelo Id
      responses:
        200:
          description: OK
          content:
            application/json: {}
        404:
          description: Not Found
          content:
            application/json: {}
    put:
      description: Editar um usuário pelo Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
            example:
              bio: "Bem vindo a minha bio"
              blog: "https://meublog.com"
              data: "2025-11-03"
              location: "Minha cidade"
              login: "bmarcelb"
              name: "Marcel Braghini"
      responses:
        200:
          description: OK
          content:
            application/json: {}
        404:
          description: Not Found
    delete:
      description: Remover um usuário pelo Id
      responses:
        200:
          description: OK
        404:
          description: Not Found

components:
  schemas:
    User:
      properties:
        login:
          type: string
        name:
          type: string
        blog:
          type: string
        location:
          type: string
        bio:
          type: string
        data:
          type: string
          format: date
          description: Data no formato yyyy-MM-dd